requirements.txt
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
streamlit
ollama
cryptography
python-dotenv
pandas
matplotlib
seaborn
jupyter
